Run swarm-client as an "unprivileged" user
The jenkins user is now added to the group that is assigned to the
Docker socket at container startup.  This is done because the GID of
that group is known to vary between systems.  A group with that GID is
added if neccesary.

The jenkins user is no longer a member of the root group as that led
to a default shell of `/sbin/halt` instead of the system's default.

The container startup now also economizes on the number of processes
that are running.